requirements.txt
pypng>=0.20220715.0
lz4>=4.3.2
lxml>=5.0.0
Pillow>=10.0.0
PyYAML>=6.0
ruamel.yaml>=0.18.0
